Add files via upload
This LoginButtonDashboard will be a button on the dashboard. When the owner clicks this button this will take the user to the scene where they get to select 1 of 2 options either "Admin Login" or "Employee Login". Which he/she can then select one and create accounts to access the application.