Add files via upload
UPDATED Report scene which will allow the user to select which one he wants to view either being "Invoice" or "Service Report". If he selects Invoice this will allow him to enter information in a form of a transaction, which then if he selects NEXT will be displayed onto a table. And same goes with Service Report button, after the user clicks this he can then enter in the top 3 services, 3 employees, and 3 customers and proceed and the info will be displayed on a final table.